add processor and aggregator services
