+++ workers/loc.api/queue/aggregator.js
@@ -6,103 +6,98 @@ const fs = require('fs')
 const unlink = promisify(fs.unlink)
 
 const {
-  uploadS3,
-  sendMail,
-  hasS3AndSendgrid,
   moveFileToLocalStorage
 } = require('./helpers')
 
-let reportService = null
-
-module.exports = async job => {
-  const aggregatorQueue = reportService.ctx.lokue_aggregator.q
-
-  try {
-    const {
-      userInfo,
-      name,
-      filePaths,
-      subParamsArr,
-      isSignatureRequired,
-      email,
-      isUnauth,
-      s3Conf,
-      emailConf,
-      language
-    } = job.data
-
-    const isEnableToSendEmail = (
-      typeof email === 'string' &&
-      await hasS3AndSendgrid(reportService)
-    )
-
-    if (isEnableToSendEmail) {
-      const s3Data = await uploadS3(
-        reportService,
-        s3Conf,
-        filePaths,
+module.exports = (
+  rootPath,
+  aggregatorQueue,
+  hasGrcService,
+  uploadToS3,
+  sendMail
+) => {
+  return async (job) => {
+    try {
+      const {
+        userInfo,
         name,
+        filePaths,
         subParamsArr,
         isSignatureRequired,
-        {
-          ...userInfo,
-          email
-        }
-      )
-
-      await sendMail(
-        reportService,
-        emailConf,
         email,
-        'email.pug',
-        s3Data.map((item, i) => ({
-          ...item,
-          isUnauth,
-          language
-        }))
+        isUnauth,
+        s3Conf,
+        emailConf,
+        language
+      } = job.data
+
+      const isEnableToSendEmail = (
+        typeof email === 'string' &&
+        await hasGrcService.hasS3AndSendgrid()
       )
 
-      for (const filePath of filePaths) {
-        await unlink(filePath)
-      }
-    } else {
-      let count = 0
-
-      for (const filePath of filePaths) {
-        await moveFileToLocalStorage(
-          reportService.ctx.rootPath,
-          filePath,
-          subParamsArr[count].name || name,
-          { ...subParamsArr[count] },
-          userInfo.username
+      if (isEnableToSendEmail) {
+        const s3Data = await uploadToS3(
+          s3Conf,
+          filePaths,
+          name,
+          subParamsArr,
+          isSignatureRequired,
+          {
+            ...userInfo,
+            email
+          }
         )
 
-        count += 1
-      }
-    }
+        await sendMail(
+          emailConf,
+          email,
+          'email.pug',
+          s3Data.map((item, i) => ({
+            ...item,
+            isUnauth,
+            language
+          }))
+        )
 
-    job.done()
-    aggregatorQueue.emit('completed')
-  } catch (err) {
-    if (err.syscall === 'unlink') {
-      aggregatorQueue.emit('error:unlink', job)
-      job.done()
-    } else {
-      try {
-        for (const filePath of job.data.filePaths) {
+        for (const filePath of filePaths) {
           await unlink(filePath)
         }
-      } catch (e) {
-        aggregatorQueue.emit('error:unlink', job)
+      } else {
+        let count = 0
+
+        for (const filePath of filePaths) {
+          await moveFileToLocalStorage(
+            rootPath,
+            filePath,
+            subParamsArr[count].name || name,
+            { ...subParamsArr[count] },
+            userInfo.username
+          )
+
+          count += 1
+        }
       }
 
-      job.done(err)
+      job.done()
+      aggregatorQueue.emit('completed')
+    } catch (err) {
+      if (err.syscall === 'unlink') {
+        aggregatorQueue.emit('error:unlink', job)
+        job.done()
+      } else {
+        try {
+          for (const filePath of job.data.filePaths) {
+            await unlink(filePath)
+          }
+        } catch (e) {
+          aggregatorQueue.emit('error:unlink', job)
+        }
+
+        job.done(err)
+      }
+
+      aggregatorQueue.emit('error:base', err, job)
     }
-
-    aggregatorQueue.emit('error:base', err, job)
   }
-}
-
-module.exports.setReportService = (rService) => {
-  reportService = rService
 }
