+++ workers/loc.api/queue/helpers/index.js
@@ -1,0 +1,15 @@
+'use strict'
+
+const getCompleteFileName = require('./get-complete-file-name')
+const {
+  moveFileToLocalStorage,
+  writableToPromise,
+  createUniqueFileName
+} = require('./utils')
+
+module.exports = {
+  moveFileToLocalStorage,
+  writableToPromise,
+  createUniqueFileName,
+  getCompleteFileName
+}
